+++ src/components/ui/MultiSelect.tsx
@@ -2,22 +2,18 @@
 import { Fragment } from 'react'
 import { Listbox, Transition } from '@headlessui/react'
 import { AiFillCaretDown } from "@react-icons/all-files/ai/AiFillCaretDown";
+import { AiOutlineCheck } from "@react-icons/all-files/ai/AiOutlineCheck";
+import {MultiSelectProps} from "@/utils/type";
 
 function classNames(...classes: string[]) {
   return classes.filter(Boolean).join(' ')
 }
 
-interface SingleSelectProps extends SelectProps {
-  value: SelectItem | null;
-  setValue: (value: SelectItem) => void;
-  label: string;
-  placeholder?: string;
-  required?: boolean;
-}
 
-export default function Select({ value, setValue, items = [], label, placeholder = "", required = false }: SingleSelectProps) {
+
+export default function MultiSelect({ values, setValues, items = [], label, placeholder = "", required = false }: MultiSelectProps) {
   return (
-    <Listbox value={value} onChange={setValue}>
+    <Listbox value={values} onChange={setValues} multiple>
       {({ open }) => (
         <div>
           <Listbox.Label className="block text-gray-700 mobile:text-sm">
@@ -25,9 +21,9 @@ export default function Select({ value, setValue, items = [], label, placeholder
             {required ? <span className="text-red-500 required-dot ml-1.5 align-middle">*</span> : <></>}
           </Listbox.Label>
           <div className="relative">
-            <Listbox.Button className="mobile:text-sm w-full min-h-[42px] cursor-default rounded-lg border-1 flex-1 appearance-none border py-2 pl-4 pr-10 text-left bg-white border-gray-300 text-gray-700 placeholder-gray-400 shadow-sm text-base focus:outline-none focus:ring-2 focus:ring-primary focus:border-transparent">
-              <span className={classNames(value ? '' : 'text-greyUnselect', 'block truncate')}>
-                {value ? value.name : placeholder}
+            <Listbox.Button className="mobile:text-sm w-full cursor-default rounded-lg border-1 flex-1 appearance-none border py-2 pl-4 pr-10 text-left bg-white border-gray-300 text-gray-700 placeholder-gray-400 shadow-sm text-base focus:outline-none focus:ring-2 focus:ring-primary focus:border-transparent">
+              <span className={classNames(values.length > 0 ? '' : 'text-greyUnselect', 'block truncate')}>
+                {values.length > 0 ? values.map((value) => value.name).join(', ') : placeholder}
               </span>
               <span className="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
                 <AiFillCaretDown className="h5 w-5 text-gray-400" aria-hidden="true" />
@@ -64,6 +60,7 @@ export default function Select({ value, setValue, items = [], label, placeholder
                               'absolute inset-y-0 right-0 flex items-center pr-4'
                             )}
                           >
+                            <AiOutlineCheck className="h-5 w-5" aria-hidden="true" />
                           </span>
                         ) : null}
                       </>
